IKZ_plugin/pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0.0",
    "setuptools-scm>=8.0",
]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
name = "ikz-plugin"
version = "0.1.0"
description = "A plugin for NOMAD containing IKZ use cases."
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Hampus Näsström", email = 'hampus.naesstroem@physik.hu-berlin.de' },
    { name = "Andrea Albino", email = 'andrea.albino@physik.hu-berlin.de' },
    { name = "Sarthak Kapoor", email = 'sarthak.kapoor@ikz-berlin.de' },
    { name = "Sebastian Brückner", email = 'sebastian.brueckner@ikz-berlin.de' },
]
license = { file = "LICENSE" }
dependencies = [
    'nomad-lab>=1.3.1',
    'nomad-material-processing',
    'nomad-measurements',
    'nomad-analysis',
    'lakeshore-nomad-plugin@git+https://github.com/IKZ-Berlin/lakeshore-nomad-plugin.git@35406777dbac424f674ad6c61b2c9bf206535859',
    'laytec_epitt_plugin@git+https://github.com/IKZ-Berlin/laytec_epitt_nomad_plugin.git@main',
]

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[project.urls]
"Homepage" = "https://github.com/FAIRmat-NFDI/AreaA-data_modeling_and_schemas/tree/main/IKZ_plugin"
"Bug Tracker" = "https://github.com/FAIRmat-NFDI/AreaA-data_modeling_and_schemas/issues"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [ 
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']

general_schema = "ikz_plugin.general:schema"
characterization_schema = "ikz_plugin.characterization:schema"
pld_schema = "ikz_plugin.pld:schema"
movpe_schema = "ikz_plugin.movpe:schema"
movpe2_growth_excel_parser = "ikz_plugin.movpe.movpe2.growth_excel:parser"
movpe1_growth_excel_parser = "ikz_plugin.movpe.movpe1.growth_excel:parser"
substrate_excel_parser = "ikz_plugin.movpe.substrate:parser"
mbe_schema = "ikz_plugin.mbe:schema"
dir_sol_schema = "ikz_plugin.directional_solidification:schema"
dir_sol_manual_protocol_excel_parser = "ikz_plugin.directional_solidification:excel_parser"
czochralski_schema = "ikz_plugin.czochralski:schema"
czochralski_multilog_parser = "ikz_plugin.czochralski:parser"


